Added full support for banner item (patterns) & items in shulker boxes
